# trabalho_paw/trabalho_paw/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PAW TP API",
    "description": "REST API para o trabalho prático de PAW",
    "contact": {
      "email": "8210455@estg.ipp.pt 8210458@estg.ipp.pt 8210461@estg.ipp.pt"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": ["http"],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["user"],
        "summary": "Logs user into the system",
        "description": "Log in the website so can use all functionalities",
        "operationId": "loginUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "item_jwt_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-access-token"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "required": [
        "nome",
        "email",
        "userName",
        "password",
        "tipoUser",
        "pontos"
      ],
      "properties": {
        "nome": {
          "type": "string",
          "example": "Pedro"
        },
        "email": {
          "type": "string",
          "example": "Pedro@estg.ipp.pt"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "example": "pedro12345"
        },
        "tipoUser": {
          "type": "string",
          "example": "Cliente"
        },
        "pontos": {
          "type": "integer",
          "format": "int",
          "example": "100"
        }
      }
    }
  }
}